DesktopSearchWeb/requirements.txt
bfastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
onnxruntime==1.16.3
opencv-python==4.8.1.78
numpy==1.24.3
pillow==10.1.0